Some thing's do)

+++ MetricsManager/DAL/Models/CpuMetrics.cs
@@ -4,7 +4,7 @@ namespace MetricsManager.DAL.Models
 {
     public class CpuMetrics
     {
-        public int AgentId { get; set; }
+        public long AgentId { get; set; }
         public int Id { get; set; }
         public int Value { get; set; }
         public long Time { get; set; }
