+++ MetricsManager/DAL/Models/DotNetMetric.cs
@@ -2,7 +2,7 @@
 {
     public class DotNetMetrics
     {
-        public int AgentId { get; set; }
+        public long AgentId { get; set; }
         public int Id { get; set; }
         public int Value { get; set; }
         public long Time { get; set; }
